Use results of obu as input for next obus
Sequenceheader was parsed, but then never used from then on.
Now, the Sequenceheader is returned, and the parser passes is to the
next obu. This prevents for example negative shit amounts, since
BitDepth would be 0 in 7.11.2 Intra prediction process.
Shifting by a negative amount is not possible according to the spec
(see Section 4.5 BItwise operators)